+++ src/api.js
@@ -56,7 +56,7 @@ class HcatApi {
      * @param {Object} queryParams - options parameter for `fetch` call
      * @returns {Promise<Object>} - server response or error
      */
-     fetchWrapper(method, endpoint, payload=null, queryParams={}, fullUrl = null) {
+     async fetchWrapper({method = "GET", endpoint = '', payload = null, queryParams = {}, fullUrl = null} = {}) {
         let options = {
             method: method,
             cors: true,