+++ helpers/index.js
@@ -6,6 +6,7 @@ module.exports = {
     createRequiredFields: require('./createRequiredFields').createRequiredFields,
     parseResponse: require('./parseResponse').parseResponse,
     setSettingValues: require('./setSettingValues').setSettingValues,
+    settleAllPromises: require('./settleAllPromises').settleAllPromises,
     wait: require('./wait').wait,
     validateTable: require('./validateTable').validateTable,
 }