+++ helpers/setSettingValues.js
@@ -1,6 +1,28 @@
 module.exports = {
-    setSettingValues: async function (setSettings, setting, values, options=false) {
-        if (options) {
+    setSettingValues: async function (setSettings, setting, values, options=false, overwrite=false, advanced=false) {
+        if (advanced) {
+            setSettings(prevState => ({
+                ...prevState,
+                "advancedConfig": {
+                    ...prevState["advancedConfig"],
+                    [setting]: {
+                        ...prevState["advancedConfig"][setting],
+                        ...values
+                    }
+                }
+            }))
+        } else if (options && overwrite) {
+            setSettings(prevState => ({
+                ...prevState,
+                [setting]: {
+                    ...prevState[setting],
+                    "options": [
+                        {"value": "", "label": ""},
+                        ...values
+                    ]
+                }
+            }));
+        } else if (options) {
             setSettings(prevState => ({
                 ...prevState,
                 [setting]: {
@@ -10,6 +32,11 @@ module.exports = {
                         ...values
                     ]
                 }
+            }));
+        } else if (overwrite) {
+            setSettings(prevState => ({
+                ...prevState,
+                [setting]: values
             }));
         } else {
             setSettings(prevState => ({